Add Strict duplications count to reconstruct command
Added ability to count only duplications that are defined by splits in
the gene tree in the recon output file. This is added as a value “X+S”
where X is a number presenting the number of duplications counted.
Also removed some leftover printf statements

Began process of being able to define a new delimiter to identifying
species names, but not yet complete.