V4.2.0 - incremental release
Introduces gene name compatibility in the “deletetrees” and
“deletetaxa” commands. This means it is possible to retain gene IDs
following filtering for various parameters as long as a delimiter
(usually ‘.’) is used to signify the end of the taxa name and the start
of the genes name in the input trees. (i.e. Human.ENSXXXXXXXX01). When
used with the “maxnamelen=delimited” modifier in the “exe command” this
will allow retention of gene names in the subsequent analyses.

Some bug fixes to the Bootstrap option in MRP and Avcon criteria which
were causing a crash with small numbers of input trees.