V4.2.1 - Introduce logging of screen output to file and command line options
Implements the “log” command that allows the users to define a file
where all output printed to screen will be echoed.
Also implemented command line options that allow the interactive mode
to be switched off and for commands to be provided with a separate
commands file. Perhaps useful for those that would like to run Clann on
a large number of datasets, using a fixed commands for instance.